Move package jeu to parent
Closes: https://github.com/nc0fr/conception-application/issues/58
