+++ src/main/java/fr/nc0/cda/modele/EtatPartie.java
@@ -4,7 +4,7 @@
  * that can be found in the COPYRIGHT file.
  */
 
-package fr.nc0.cda.modele.jeu;
+package fr.nc0.cda.modele;
 
 /** Représente les états possibles d'une partie à deux joueurs. */
 public enum EtatPartie {
