+++ src/main/java/fr/nc0/cda/modele/joueur/StrategieAiNimAleatoire.java
@@ -6,8 +6,8 @@
 
 package fr.nc0.cda.modele.joueur;
 
-import fr.nc0.cda.modele.jeu.Choix;
-import fr.nc0.cda.modele.jeu.Plateau;
+import fr.nc0.cda.modele.Choix;
+import fr.nc0.cda.modele.Plateau;
 import fr.nc0.cda.modele.nim.ChoixNim;
 import fr.nc0.cda.modele.nim.PlateauNim;
 import fr.nc0.cda.vue.Ihm;
